feat: replaced piptools with poetry for easier use, updated readme
Readme now contains basic instructions on how to use the project and where to download specific files for use by the program.
