+++ .gitignore
@@ -159,6 +159,10 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 #.idea/
 
+# Poetry
+poetry.lock
+
+# Project specific
 Streetcheck Output/
 streetcheck_storage/
 tests/streetcheck_test_csvs
