--- requirements-dev.in
@@ -1,8 +1,0 @@
-mypy
-ruff
-black
-isort
-types-requests
-types-tqdm
-pytest
-matplotlib
