--- requirements-dev.txt
@@ -1,146 +1,0 @@
-#
-# This file is autogenerated by pip-compile with Python 3.11
-# by the following command:
-#
-#    pip-compile --extra=dev --output-file=requirements-dev.txt pyproject.toml
-#
-attrs==23.2.0
-    # via fiona
-black==23.7.0
-    # via ukconstuencystreetcheck (pyproject.toml)
-certifi==2023.7.22
-    # via
-    #   fiona
-    #   pyogrio
-    #   pyproj
-    #   requests
-charset-normalizer==3.2.0
-    # via requests
-click==8.1.6
-    # via
-    #   black
-    #   click-plugins
-    #   cligj
-    #   fiona
-click-plugins==1.1.1
-    # via fiona
-cligj==0.7.2
-    # via fiona
-contourpy==1.2.1
-    # via matplotlib
-cycler==0.12.1
-    # via matplotlib
-cython==3.0.0
-    # via ukconstuencystreetcheck (pyproject.toml)
-et-xmlfile==1.1.0
-    # via openpyxl
-fiona==1.9.6
-    # via geopandas
-fonttools==4.50.0
-    # via matplotlib
-geojson==3.1.0
-    # via ukconstuencystreetcheck (pyproject.toml)
-geopandas==0.14.3
-    # via ukconstuencystreetcheck (pyproject.toml)
-greenlet==2.0.2
-    # via sqlalchemy
-idna==3.4
-    # via requests
-iniconfig==2.0.0
-    # via pytest
-isort==5.12.0
-    # via ukconstuencystreetcheck (pyproject.toml)
-joblib==1.3.2
-    # via scikit-learn
-kiwisolver==1.4.5
-    # via matplotlib
-matplotlib==3.8.4
-    # via ukconstuencystreetcheck (pyproject.toml)
-mypy==1.5.0
-    # via ukconstuencystreetcheck (pyproject.toml)
-mypy-extensions==1.0.0
-    # via
-    #   black
-    #   mypy
-numpy==1.25.2
-    # via
-    #   contourpy
-    #   matplotlib
-    #   pandas
-    #   pyogrio
-    #   scikit-learn
-    #   scipy
-    #   shapely
-openpyxl==3.1.2
-    # via ukconstuencystreetcheck (pyproject.toml)
-packaging==23.1
-    # via
-    #   black
-    #   geopandas
-    #   matplotlib
-    #   pyogrio
-    #   pytest
-pandas==2.0.3
-    # via
-    #   geopandas
-    #   ukconstuencystreetcheck (pyproject.toml)
-pathspec==0.11.2
-    # via black
-pillow==10.3.0
-    # via matplotlib
-platformdirs==3.10.0
-    # via black
-pluggy==1.4.0
-    # via pytest
-pyogrio==0.7.2
-    # via ukconstuencystreetcheck (pyproject.toml)
-pyparsing==3.1.2
-    # via matplotlib
-pyproj==3.6.1
-    # via geopandas
-pytest==8.1.1
-    # via ukconstuencystreetcheck (pyproject.toml)
-python-dateutil==2.8.2
-    # via
-    #   matplotlib
-    #   pandas
-pytz==2023.3
-    # via pandas
-requests==2.31.0
-    # via ukconstuencystreetcheck (pyproject.toml)
-ruff==0.0.284
-    # via ukconstuencystreetcheck (pyproject.toml)
-scikit-learn==1.4.1.post1
-    # via ukconstuencystreetcheck (pyproject.toml)
-scipy==1.13.0
-    # via scikit-learn
-shapely==2.0.3
-    # via
-    #   geopandas
-    #   ukconstuencystreetcheck (pyproject.toml)
-six==1.16.0
-    # via
-    #   fiona
-    #   python-dateutil
-sqlalchemy==2.0.19
-    # via ukconstuencystreetcheck (pyproject.toml)
-threadpoolctl==3.4.0
-    # via scikit-learn
-tqdm==4.66.1
-    # via ukconstuencystreetcheck (pyproject.toml)
-types-requests==2.31.0.2
-    # via ukconstuencystreetcheck (pyproject.toml)
-types-tqdm==4.66.0.20240106
-    # via ukconstuencystreetcheck (pyproject.toml)
-types-urllib3==1.26.25.14
-    # via types-requests
-typing-extensions==4.7.1
-    # via
-    #   mypy
-    #   sqlalchemy
-tzdata==2023.3
-    # via pandas
-urllib3==2.0.4
-    # via requests
-uvloop==0.19.0
-    # via ukconstuencystreetcheck (pyproject.toml)
