--- requirements.in
@@ -1,12 +1,0 @@
-tqdm
-cython
-SQLAlchemy
-pandas
-requests
-uvloop
-scikit-learn
-geojson
-shapely
-geopandas
-pyogrio
-openpyxl
