+++ ukconstituencystreetbystreet/__main__.py
@@ -1,4 +1,4 @@
-from ukconstituencyaddr import outputter
+from ukconstituencystreetbystreet import outputter
 
 if __name__ == "__main__":
     outputter.output_csvs()
