+++ ukconstituencystreetbystreet/ons_constituencies.py
@@ -11,9 +11,9 @@ from typing import Dict, Optional
 import pandas as pd
 from sqlalchemy.orm import Session
 
-from ukconstituencyaddr import config
-from ukconstituencyaddr.db import cacher
-from ukconstituencyaddr.db import db_repr_sqlite as db_repr
+from ukconstituencystreetbystreet import config
+from ukconstituencystreetbystreet.db import cacher
+from ukconstituencystreetbystreet.db import db_repr_sqlite as db_repr
 
 logger = logging.getLogger(__name__)
