+++ ukconstituencystreetbystreet/ons_oa.py
@@ -20,10 +20,10 @@ import numpy as np
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
-from ukconstituencyaddr import ons_constituencies
-from ukconstituencyaddr import config
-from ukconstituencyaddr.db import cacher
-from ukconstituencyaddr.db import db_repr_sqlite as db_repr
+from ukconstituencystreetbystreet import ons_constituencies
+from ukconstituencystreetbystreet import config
+from ukconstituencystreetbystreet.db import cacher
+from ukconstituencystreetbystreet.db import db_repr_sqlite as db_repr
 
 
 class OnsOaField(enum.StrEnum):
